Rename access API model to auth API model
