+++ st2auth/st2auth/controllers/access.py
@@ -22,7 +22,7 @@ from oslo.config import cfg
 
 from st2common.exceptions.auth import TTLTooLargeException
 from st2common.models.api.base import jsexpose
-from st2common.models.api.access import TokenAPI
+from st2common.models.api.auth import TokenAPI
 from st2common.services.access import create_token
 from st2common import log as logging
 from st2auth.backends import get_backend_instance
