+++ st2common/st2common/models/api/auth.py
